Move common objects to common code.
