+++ hyper/common/bufsocket.py
