+++ hyper/http20/exceptions.py
@@ -5,20 +5,6 @@ hyper/http20/exceptions
 
 This defines exceptions used in the HTTP/2 portion of hyper.
 """
-class SocketError(Exception):
-    """
-    An error occurred during socket operation.
-    """
-    pass
-
-
-class LineTooLongError(Exception):
-    """
-    An attempt to read a line from a socket failed because no newline was
-    found.
-    """
-    pass
-
 
 class HTTP20Error(Exception):
     """
@@ -54,13 +40,3 @@ class ProtocolError(HTTP20Error):
     The remote party violated the HTTP/2 protocol.
     """
     pass
-
-
-# Create our own ConnectionResetError.
-try:  # pragma: no cover
-    ConnectionResetError = ConnectionResetError
-except NameError:  # pragma: no cover
-    class ConnectionResetError(Exception):
-        """
-        A HTTP/2 connection was unexpectedly reset.
-        """
